+++ packages/docs/src/stories/Box.stories.tsx
@@ -10,8 +10,27 @@ export default {
         <Text>Testando o elemento Box</Text>
       </>
     ),
+    variant: 'secondary',
+    hasBorder: true,
   },
-  argTypes: {},
+  argTypes: {
+    variant: {
+      options: ['primary', 'secondary'],
+      control: {
+        type: 'inline-radio',
+      },
+    },
+    hasBorder: { control: { type: 'boolean' } },
+  },
 } as Meta<BoxProps>
 
-export const Primary: StoryObj<BoxProps> = {}
+export const Primary: StoryObj<BoxProps> = {
+  args: {
+    variant: 'primary',
+  },
+}
+export const Secondary: StoryObj<BoxProps> = {
+  args: {
+    variant: 'secondary',
+  },
+}